Moved the routes around so that it's easier to organize
