+++ Sources/App/Collections/V1Collection.swift
@@ -21,19 +21,8 @@ public final class V1Collection: RouteCollection {
     // MARK: - User Controller
     let userController = UserController(view)
     
-    // web
-    builder.group("users") { user in
-      user.group(AuthenticateMiddleware(), AdminMiddleware()) { auth in
-        auth.get("/", handler: userController.index)
-        auth.post("/", handler: userController.store)
-        auth.patch(":userId", handler: userController.update)
-        auth.delete(":userId", handler: userController.destroy)
-      }
-      user.get("login", handler: userController.loginView)
-      user.post("login", handler: userController.loginWebPost)
-    }
     api.post("auth", "login", handler: userController.loginAPIPost)
-    
+
     // api
     api.grouped(AuthenticateMiddleware(), AdminMiddleware()).group("users") { user in
       user.get("/", handler: userController.index)
@@ -48,15 +37,8 @@ public final class V1Collection: RouteCollection {
     let eventController = EventController(view)
     
     // MARK: Events API
-    let eventsAPI = api.grouped("events")
-    eventsAPI.get("/", handler: eventController.all)
-    
-    // MARK: Events Web View
-    let events = builder.grouped("events")
-    events.get("/", handler: eventController.index)
-  
-    // MARK: - HomeController
-    let homeController = HomeController(view)
-    builder.get("/", handler: homeController.index)
+    api.group("events") { event in
+      event.get("/", handler: eventController.all)
+    }
   }
 }
