+++ Sources/App/Setup/Config+Setup.swift
@@ -29,8 +29,9 @@ extension Config {
   /// Add all models that should have their
   /// schemas prepared before the app boots
   private func setupPreparations() throws {
-    preparations.append(EventType.self)
+    preparations.append(UpdateEventTable.self)
     // no migrations, just regular here
+    preparations.append(EventType.self)
     preparations.append(Event.self)
     preparations.append(Asset.self)
     preparations.append(Pivot<Event, Asset>.self)
